ocr_back/requirements.txt
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
django
django-cors-headers
djangorestframework